+++ CHANGELOG.md
@@ -5,14 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
+- [\#53](https://github.com/openzklib/openzl/pull/53) Add Initial Halo2 Plugin
 - [\#37](https://github.com/openzklib/openzl/pull/37) Add BitDecomposition in eclair
 - [\#47](https://github.com/openzklib/openzl/pull/47) Migrate docs to main OpenZL repo
 - [\#36](https://github.com/openzklib/openzl/pull/36) Finish OpenZL migration feature requirements
 - [\#34](https://github.com/openzklib/openzl/pull/34) Migrate Poseidon from the Manta-Network codebase
 - [\#4](https://github.com/openzklib/openzl/pull/4) Add CI and Relevant Contribution Files
 - [\#3](https://github.com/openzklib/openzl/pull/3) Migrate some of OpenZL from the Manta-Network codebase
-- [\#4](https://github.com/openzklib/openzl/pull/4) Add CI and Relevant Contribution Files
-- [\#34](https://github.com/openzklib/openzl/pull/34) Migrate Poseidon from the Manta-Network codebase
 
 ### Changed